Look for the glade files in multiple places
During development we just load the glade files from the current
directory. However, in production, the glade files will want to be
stored away from the binary in some suitable location, say,
/usr/share/auric/

So look for the glade files in the current working directory, if not,
then look in /usr/share/auric/ which is where the rpm will install them.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>