Expand use of view invoice details
This was originally used for viewing the repeat invoices. Genericise
this so that it can also be used to view invoices for a selected entity
from the entities tab.

The bulk of the change is actually just renaming the auric_ri_vid.glade
file and renaming ri_vid to vid.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>